moved regex responsibility out of route resolver
