+++ src/utilities/index.ts
@@ -1,8 +1,11 @@
 export * from './createRouter'
 export * from './mergeParams'
+export * from './paramValidation'
 export * from './params'
 export * from './paramsFinder'
-export * from './paramValidation'
 export * from './path'
 export * from './random'
+export * from './routeRegex'
 export * from './routeResolver'
+export * from './string'
+export * from './urlAssembly'
