+++ src/utilities/routeResolver.ts
@@ -33,18 +33,9 @@ export function resolveRoutes(routes: Routes, parentContext: ParentContext = {})
         name: route.name,
         path: fullPath,
         params: fullParams,
-        regex: generateRouteRegexPattern(fullPath),
       })
     }
 
     return value
   }, [])
-}
-
-export function generateRouteRegexPattern(value: string): RegExp {
-  const optionalParamRegex = /(:\?[\w]+)(?=\W|$)/g
-  const requiredParamRegex = /(:[\w]+)(?=\W|$)/g
-
-  const routeRegex = value.replace(optionalParamRegex, '(.*)').replace(requiredParamRegex, '(.+)')
-  return new RegExp(`^${routeRegex}$`)
 }
